file update
laravel breeze updated...the tailwindcss install isn't working right for the login and reg pages so i included the cdn in the app and guest layout files to get it working. I'm not worried about it breaking in the future because i do not intend to revert the changes as this is a demo project.